The LAST compiler warning :P

+++ src/server/sv_send.c
@@ -425,7 +425,7 @@ static _Bool Sv_RateDrop(sv_client_t *cl) {
 
 	size_t total = 0;
 
-	for (int32_t i = 0; i < lengthof(cl->frame_size); i++) {
+	for (size_t i = 0; i < lengthof(cl->frame_size); i++) {
 		total += cl->frame_size[(sv.frame_num - i) % lengthof(cl->frame_size)];
 		if (total > cl->rate) {
 			cl->surpress_count++;
